[LIGO-208] Make Standard scopes more robust
Problem: Currently, if FromCompiler scopes fail during scoping, it will
fail altogether. This is particularly important for Standard scopes,
since we want to have as much information from the compiler as possible.

Solution: Add to `HasScopeForest` a new method to scope only a single
contract and refactor FromCompiler and Fallback to use the default
implementation for a graph and provide an implementation for a single
file. Now Standard has fine control over the results of both Fallback
and FromCompiler, and we can take a decision: if both scopes succeed,
merge them, otherwise use Fallback. Note that FromCompiler will still
fail altogether if only one of them fail, but this is not important
right now, because we care about the result of Standard scopes.

Note that I chose to keep the `scopeForest` as a method, in case in the
future we want to refactor it for performance or something, but it can
also be refactored as a function instead.