[LIGO-208] Add tests
Problem: We should test whether our changes work as intended.

Solution: Add a directory with a good and a bad file. Standard scopes
should be able to scope the "good" file even if the "bad" file fails
(where it should use only Fallback).

I uncommented some tests along the way, and provided better issue
numbers for them.

Also, there was a problem in diagnostic tests, since parser errors are
actually part of the parsing system and not fallback scopes, they should
be separated so that we can get proper test results.