[LIGO-591] Release plugin resources
Problem: unlike in morley-debugger, here in `dispose` method we do
nothing.
This is not good, as plugin keeps a server and debug adapter threads,
and these are resources that must be released.

Solution: move `server` and `client` to module-local variable (instead
of method-local), close server with all our resources on debug session
termination.