[LIGO-575] Improve CameLIGO type highlighting
Problem: We want better type highlighting for CameLIGO in VS Code, since
there is none currently.

Solution: Create new regexes for various constructions in CameLIGO and
add add them to the generator. I tried to put the corresponding things
to Vim and Emacs as well, and it seems to work somewhat well for Vim,
although Emacs has very limited support so the changes didn't affect it
too much.

A new `Type_var` was added to support polymorphic type variables.

A few bugs were fixed on the way:
* We can't reference patterns that are added later in `Textmate.ml` such
  as `string`. Add a check if we're using a "common" thing.
* If we reference the grammar with `$self`, it will replace it with
  `#$self`. Don't do any transformations in this case. Turns out that we
  didn't use this feature, but it's nice to get it fixed anyway.
* `identifierconstructor` expects one or more spaces at the end. Use a
  word boundary instead.
* `let_binding_match3` and `let_binding_match2_ligo` forgot to include
  word boundaries for Vim.