[LIGO-550] Add backend changes
Problem: During the prototype, we created the backend in a sort of
makeshift way, without proper review or commit messages. We need to
commit the backend to the master source tree with proper review.

Solution: Add all the backend changes to a single commit. The changes
introduced by this commit are
* a backend that can compile LIGO contracts of all three dialects.
* a (small) test suite
* a README for building and running the backend