[LIGO-318] Create a skeleton for the scoping algo
Problem: The scoping algorithm needs to walk the syntax tree collecting
scopes and recording references. It needs to be able to handle all
possible syntax constructions.

Solution: Set up a skeleton for the scoping algorithm with placeholders
for every possible type of node in the syntax tree. We maintain both a
`StateT (Map DeclRef ScopedDecl)` to track all variables and their
references in the code, as well as a `ReaderT (HashMap Text Range)` for
the variables that are in-scope at a given point. We walk through the
syntax tree and use these two containers to determine the scope for each
variable, as well as collect all of its references. At the end, we
return a ScopeForest that contains all this information about scopes and
references. See the issue description for LIGO-318 for a more detailed
description of the algorithm.