[LIGO-575] Improve JsLIGO type highlighting
Problem: We want better type highlighting for JsLIGO in VS Code, since
there is none currently.

Solution: Create new regexes for various constructions in JsLIGO and add
them to the generator. I tried to put most of the corresponding things
to Vim and Emacs as well, but JsLIGO doesn't really support Vim and
Emacs, so the changes aren't visible in there.