[LIGO-575] Fix object/record expression highlighting
Problem: The syntax for object/record expressions and types are the
same, so the syntax highlighting ends up highlighting these expressions
as if they were types.

Solution: Create a new scope for object/record expressions. Now when the
syntax highlighting see an opening brace, it knows it will have to deal
with expressions instead of types (which will be nested in a type
annotation or a type definition). Add properties/fields expressions so
that the syntax highlighting doesn't confuse it with types.