[LIGO-610] Create a better structure for messages
Problem: We only use a plain `Text` to identify errors. We should have
something else to understand the cause of the diagnostic.

Solution: Create a `MessageDetail` type for diagnostics that allows to
better specify the origin of the message. Move things from `Parser` to
`Diagnostic` so we can reuse it without cyclic imports (and also to make
it a bit more organized).