[LIGO-639] Check for invalid UTF-8 strings.
Problem: The language server crashes when there is a LIGO file with
invalid UTF-8 during the indexing.

Solution: Try to stricly decode the UTF-8 and return it if there are no
errors. Otherwise, use `lenientDecode` to substitute invalid sequences
with the Unicode replacement character U+FFFD (�) and log an error
message to the terminal.

Also, we previously used `Text.IO.readFile`, but it seems to use
system-dependent encoding and have poor performance. See this blog post:
https://www.snoyman.com/blog/2016/12/beware-of-readfile/

One side effect: If we open a file with an invalid encoding (e.g.:
UTF-16), then it will report no errors, as this file will come directly
from LSP. However, if we open a file that was already indexed, then
there will be a few "Unexpected" and "Unrecognized" errors in it.