[LIGO-594][LIGO-579] Adapt debugger code for new `morley-debugger` version
Problem: I've supported pausing on breakpoints in different contracts
in `morley-debugger`. If we want to use this feature we need to change
`morley-debugger` version and adapt LIGO debugger code for these
changes.

Solution: change commit hash in `stack.yaml` for `morley-debugger`.
Add some fixups for typechecking.

:NOTE: hash in `stack.yaml` points to opened MR. We should change
it to `master` commit hash when this MR is merged.