[LIGO-203] Fix failing tests
Problem: The LIGO team created a `.ligoproject` on the root directory,
but now our `lsp-handlers-test` tries to index the entire LIGO directory
for each test.

The `preprocessor` directory was removed and caused tests to fail.

In addition, one new failing contract was added.

Solution: Add `.ligoproject` to restrict tests as much as possible, so
it doesn't look outside of these directories. Turns out the `LIGO_ENV`
flag for tests is now useless, so we can remove the code that checks for
this since there will always be a `.ligoproject` file.

Remove tests for `preprocessor`.

Add failing contract to ignore list.