[LIGO-620] Fix executeCommand output
Problem: executeCommand provides to many different
return options. It can return undefined and print
message to the console, return undefined without printing
message or return string value. We need to return something
general in every possible situation.

Solution: Return value of new ExecutionResult type,
which provides information about what happened while
execution, and possibly the result, if it exists.
Also changed arguments a bit, so now you can decide
wether you want output to be printed to the console,
or to the file. Note, since in case of printing to file
the ligo binary is used, you can't get output to the
console in this case, because binary itself do not produce it.
Anyway the result and the status variable will be returned.