[LIGO-611] Properly concat whitespaces in Vim
Problem: In Vim, we can use `skipempty skipwhite` to skip empty lines
(and newlines) and whitespace when there are `nextgroup`s to not break
syntax highlighting regardless of whitespace.

Solution: Make VSCode and Emacs implicitly insert spaces when there are
multiple rules in a list. Make Vim use `skipempty skipwhite` instead.
Fix regexes to properly deal with whitespace and remove the explicit
skipping. Note that this aids in Vim, but Emacs (I think?) and VS Code
can only match one line at a time. Good enough!

By the way, ask @heitor.toledo if you want whitespacing skipping on a
per-rule setting. I originally did this, but for simplicitly, stashed
the work and opted for a simpler alternative of always inserting it.