[LIGO-611] Make types not break comments and a few improvements
Problem: When there is a comment within a type, it will break, causing
ugly highlighting. This is because comments can only appear at top
level, and a `typeparentheses` match before the block comments.

Solution: Put comments last so they are always matched regardless of
existing rules. While we're here, also turn on spell checking for
comments and strings.