[LIGO-611] Use @top instead of $self in Vim
Problem: $self is a VSCode thing, in Vim you are supposed to use TOP.

Solution: Create a syntax cluster called `top` and alias it to contain
just `TOP`. `TOP` in Vim represents every syntax rule which isn't
`contained`. The reason I use `@top` (to refer to this cluster) and not
`TOP` is because in Vim, `TOP` must be the first in a `contains` or
`containedin`, and in addition, `TOP,x,y,z` means "TOP except x, y, and
z" which is unintuitive.

I also took the liberty to simplify a few regexes by removing unneeded
parentheses and disallowing to match the empty string (it may cause
performance issues).