[LIGO-648] Track which files are dirty
Problem: We need to know which files are dirty or not to be able to know
which ones we need to save to the disk and which we don't.

Solution: Make `reOpenDocs` keep track of information regarding the
currently open documents (currently only whether it's dirty or not).
I changed it to a `StmMap` to make it work a bit better with
concurrency. Finally, if a document is changed, mark it as dirty, and if
it's saved, mark it as not dirty.