[LIGO-627] Let the user specify an entrypoint
Problem: We rely on either a hardcoded entrypoint being set in
`package.json` or if none is provided, defaulting to `main`. But for
better UX, we want the user to provide an entrypoint from a box.

Solution: The TypeScript part had to be refactored a bit to not care a
lot about the initialization order, this was achieved by changing the UI
to accept a function that saves something in the server as soon as we
get an input from the user.
A method was added to save the picked file, as well as the inputed
entrypoint. The backend now has a few `Maybe`s as we have an
initialization order for things, and each handled should set these
things as they are initialized, but unfortunately the order of
initialization steps are still pretty dependent on how the handlers are
called.