[LIGO-627] Clean up the code
Problem: We should refactor so that `GetContractMetadata` has no side
effects.

Solution: Create a new `CompileContract` that finally compiles the
contract on the backend. I also took the liberty of creating a few
helpers for extracting things from the language server state, as there
were a lot of boilerplate and repeated code.