[LIGO-648] Only write files that are dirty
Problem: We should only write to the disk files that are dirty, but we
also write clean ones.

Solution: Create a new `DirtyStatus` that keeps track of whether a file
is clean or dirty. Use it for `withLigo` and propagate changes. Now, in
RIO, we use the info of whether the file is dirty or not to properly
insert it in `Dirty` or `Clean`.