[LIGO-684] Log error on decoding failure
Problem: The language server dies if LIGO provides a wrong scope.

Solution: Skip the failed scope, and continue parsing if possible. As
for the ignored scope, log an error with more information.

I'm not sure if this is the best approach, we likely can also return it
as a range, since we have the location of the scope where it happened.