[LIGO-681] Fix attributes not recognized
Problem: %Michelson is not the only code injection, as LIGO also uses
%external. To be honest, any code injection may appear.

Solution: Generalize the code to deal not only with Michelson, but any
arbitrary code injection.

Note that [%lang should appear as one token, but tree-sitter is bugged
and so we're accepting spaces (but we could also workaround on Haskell's
side to support this).