[LIGO-645] implement `documentHighlight` method in lsp server
Problem: this method enable clients to highlight not all text occurrences of the
name of variable (default behavior in vscode) but only related references. It's
important feature because of prevalence name hiding and sameness of type and value
names in functional dialects.

Solution: implement the method by reusing implementation of `FindReferences`