[LIGO-318] Fix Capture to take one accessor
Problem: According to the LIGO grammars, a record capture can only
contain a single accessor, not a sequence of nested accessors. But in
AST.Skeleton, we define `Capture [it]`, which allows Capture to take
multiple accessors.

Solution: Change the definition of Capture in AST.Skeleton to
`Capture it`, and update the codebase accordingly.