[LIGO-318] Flesh out the walk logic
Problem: We need to fill out the scoping algorithm with the logic for
walking each node in the syntax tree.

Solution: Add the logic for walking each node in the syntax tree. Nodes
that generate new scopes, or whose children may generate scopes, yield a
Just (ScopeTree, [DeclRef]) that contains information on the generated
scopes. Nodes that don't generate any scopes yield Nothing, but we still
walk them to collect references.