[LIGO-681] Fix highlighting for JsLIGO attributes
Problem: JsLIGO has a conflict with attributes and comments (as
attributes are inside comments).

Solution: Allow matching "extra regions" inside comments and strings.
For JsLIGO, embed attributes and remove it from the top level repo.

Note that our solution highlights everything that looks like an
attribute inside a comment. In reality, for a comment to be considered
an attribute, the first thing that needs to appear inside the comment is
a @. Emacs support for the new feature is missing.

Good enough for now.