[LIGO-699] Use exceptions to report errors
Problem: If anything happens in extension, it reports
about it returning undefined, which in turn should
be handled up the stack. With extending capabilities of
extension, the number of such checks grows enormously,
Moreover, we report about such findings, using different
methods, and different formats, and sometimes, simply
push returned undefined upper the stack. All of this
makes codebase ugly and inconvenient.

Solution: Now if anything bad happens, immediately
an exception is thrown, which cancels current evalution, and
is caught only in the call function. Afterwards it is
reported using only one format.