[LIGO-688] Refactor TextMate and VSCode generators
Problem: We change Emacs and Vim to take a list of syntaxes, but Emacs
and VS Code still expect individual arguments, resulting in a lot of
duplicated code. Moreover, these two generators are almost identical,
with a lot of duplication.

Solution: Extract common functionality into a new function and fold over
the list of syntaxes, stopping at the first `Error`, if any. This should
simulate a `traverse` in Haskell.