[LIGO-698] Use persistent LIGO processes
Problem: The LIGO team added a new `ligo daemon` CLI that allows to keep
LIGO open as a persistent process, and we should make use of it in the
language server.

Solution: Add functions to start, stop, and restart the daemon. Now, we
save the LIGO processes in a `Pool` in `HasLigoClient` and reuse them.

In the language server, whenever the configuration changes, or the
indexing directory changes, we update these settings.

I also added a few debugging utilities to help with testing the new
feature.

Note: There is a caveat. We read the input stream character by
character. Likely we want to find a more efficient way in the future!