[LIGO-720] Bump stack lts to 19.16 + adapt code to new morley-debugger
Problem: recently `lsp` was bumped to lts-19.16. Since we want to
have a dependency on it we should also bump our lts

Solution: change resolver in `stack.yaml` to lts-19.16.
While bumping to new lts I had to change `morley-debugger` version
to the newest one. All changes in `Snapshots` and `Impl` modules
are copypasted from `morley-debugger`