[LIGO-559] Show operations and storage on fine-termination
Problem: usually we want to see the result of contract execution
after debugging. However, at this moment we can see it only in last
snapshot and we can skip it by the accident

Solution: print final list with operations and new storage in debug console.
Also print the old storage to compare it with the new one