[LIGO-218] Handle missing field declaration
Problem: In a type declaration, we forgot to handle the case when we
have a field declaration. This is necessary so tests pass in the next
commit, but it's not complete; there are still missing cases for type
declarations that must be resolved in a future issue.

Solution: Besides `Variant`, check if we got a `TField` for a type
declaration and appropriately scope it.