[LIGO-218] Use references when looking for the env
Problem: Our current `lookupEnv` function only checks for a name in the
environment, but this will quickly break once we introduce modules.
Supposing that we have `foo` in the top-level scope and we try to use
`Foo.foo` (and supposing that it was not declared), then it'll find
`foo` instead, which is incorrect.

Solution: When looking for references, ensure that whatever we are
looking for is also the origin of a declaration or a reference of a
declaration. With this, we can be more certain that whatever we're
looking for is more correct, and not some shadowed declaration within
the file.