[LIGO-218] Support modules in scopes
Problem: We want to support modules and module fields in our scopes, as
well as all of their references. Currently, they aren't walked on or
declared.

Solution: First, carry the scope environment in a new record which
remembers the current namespace that we're scoping. Also change whatever
is in scope to remember not only its name, but also its declaration
namespace. `DeclRef` and `ScopedDecl` also remembers their namespaces
now.

Now, whenever we enter a `BModuleDecl` or a `ModuleAccess`, we can add
its corresponding modules as we traverse the tree `local`ly, so that
`walk`ing on them will lookup the correct scope. This avoids, for
example, having an `add` and a `A.add` being considered the same thing.

To search for modules, we need to have a notion of level. I decided to
just add a new `ModuleLevel`, since it's not something that happens at
either term or type levels. With it, our `findScopedDecl` will work
properly.

On the way, I also made a few small changes:
* Use `::` rather than `.` for printing module access, to disambiguate
  with field access.
* When pretty-printing a `sexpr`, avoid adding a trailing whitespace
  when the given list is empty.
* Other small changes.

There is still work to do to fully support modules in scopes, namely:
* Completion of module fields.
* Properly pretty-printing namespaces.
* Resolving references from module aliases.