[LIGO-726] Remove chain related part from pm
Problem: EthereumStudio as a modular project very well
splitted components related to chain and base components.
It is good for them but make development a little bit confused.

Solution: as we are going to use IDE only for one chain it is
not needed to use such segmentation. We can choose different
way to logically separate project manager and project settings
later, but right now simply collect all implementations in one
module.