[LIGO-722] Implement statement visiting
Problem: at this moment we're stopping only on expressions. However,
if we want to implement step granularity, we need to have stopping
at statements

Solution: each time we face an instruction with location meta we should
find all unrecorded statements. Statement is unrecorded if we don't have
its range stored or if it is present in some loop. Statements are:
1. let-ins
2. variable assignments
3. function calls inside some block structure
(e.g. `begin ... end` in `CameLIGO` or `{}` in `JsLIGO`)