[LIGO-218] Support module aliases
Problem: We now support module declarations, but we must also support
module aliases.

Solution: Supporting aliases require a bit more invasive changes, as we
previously didn't have anything doing this kind of scope resolution*.
First, we need to keep track of what is an actual reference, and what is
an alias, and to what this alias resolves to.

Because of the way Fallback scopes work, I also had to change
ExtendedRefKind to remember whether it's an alias or an ordinary
reference, otherwise we'd lose this information along the way and we
wouldn't be able to retrieve it again.

In Fallback, we now can keep track of what is a module alias, by
checking the reference kind in the in-scope map. However, we must also
return what were our declared aliases in order to preserve this
information.

We now handle the BModuleAlias case, inserting it as an alias reference.
When we walk over a module path, we can resolve the aliased name at each
step in order to find the original declaration.

While merging scopes, we also need to properly handle namespaces, and
use the ones from Fallback, since LIGO doesn't yet output module
information.

* Well, in the completion capability, we actually resolve types in order
  to find field declarations. We could do it in a per-capability basis,
  but let's do it in Fallback to avoid duplicate work!