[LIGO-218] Improve pretty printing for modules
Problem: When hovering at a definition, we sometimes see `(:: A) titi`
instead of `A.titi`. Also, modules appear as `A : A`.

Solution: First, in our `ScopedDecl` parser, make sure we carry
`dialect` all the way down to the printers by using a `Reader`. This
way, we print the correct dialect and not the S-Expressions.

Then, in `Hover`, the first problem is automatically resolved. For the
second problem, if we have a module, simply don't try to print its
"type", since there is none.