[LIGO-218] Lookup for decls in previous namespaces
Problem: Suppose something like this:

```
module M = struct
  let x : int = 0
  module N = struct
    let y : int = x
  end
end
```

In this situation, if we try to find `x` within `y`, it will attempt to
use `M.N.x`, but its namespace is actually `M.x`.

Solution: Lookup for declarations in previous namespaces. Note that we
cannot do so blindly. Consider the following scenario:

```
let b : int = 0
let a : int = X.b
```

If we attempt to look for `X.b`, and, supposing that it's not in-scope,
it would look for `b` instead, which is incorrect. Because of this, we
also keep track of whether the current namespace is qualified or
unqualified and do the backwards search only if it's unqualified.