[LIGO-677] Add generate-deploy-script to backend
Problem: We need to be able to generate the deploy script entirely on
the backend.

Solution: Add generate-deploy-script to the API. We implement
generateDeployScript using the deploy script logic in the old webide as
a starting point. The script requires an estimate of the burn fee, so we
generate the estimate by performing a dry-run of the Michelson contract.