[LIGO-391] Resolve decoding issues after LIGO 0.51.0
Problem: LIGO 0.51.0 was released, but with it, some tests started to
fail.

Solution: It seems the `wrap_content` in `LigoTyBinder` became optional
(or perhaps removed?), and `type` in `LigoTypeForAll` as well, so put a
`Maybe` there.

Also, I think there was a mistake in `LigoTypeForAll`, we skip 3 letters
but it should be 4 instead, so fixed.

I don't know why JsLIGO arrow type tests are failing now, but I created
LIGO-797 to deal with this issue.