[LIGO-700] fix bug with reference finding
Problem: there is no `ScopedDecl` values in the `NameDecl name` AST node that is
related to the scope of the `name`. It's a corner case where we have to know
about the scope in place that is not covered by it.

Solution: on the phase of building AST tree from `ScopeForest` we add one
additional scope to `NameDecl`.