[Chore] Rewrite case on switch button
Motivation: this change is arguable, but re-writing that case to pattern
match on `undefined` (instead of having a special `if`) seems simpler
and reports the intention more clearly.

It is also safer, as now type system forces us to update this code if a
new type of switch button is added.