[LIGO-802] Move API, SwaggeredAPI to Api module
Problem: API and SwaggeredAPI are purely declarative types, describing
the topology of the backend API without specifying any implementation
details. Therefore, they are not appropriate for the Server module,
which is responsible for setting up the actual implementation of the
backend server.

Solution: Move API and SwaggeredAPI to the Api module.