[Chore] Change decoding error
Problem: currently when for `ligo compile contract` JSON parsing fails,
we print only the output of ligo and do not mention the reason.

This is something what we cannot afford as the output may be extremly
large. And for debugging purposes we want to actually see the reason of
error.

Not like this is what the user actually want to see, but if we approach
it from the point of view of the user, then we have to record both the
info for us in one channel and info for the user into another channel.
Let's handle this later.
For now it's not so important because we can always reproduce what user
does assuming that they know the version of `ligo` they use.

Solution: change decoding to prompt only the reason of failure, not the
`ligo`'s output.