[LIGO-203] Make a test work
Problem: there is a corner case in `ligo` that stopped working.
This makes one test fail.

Solution:

Create a ticket in ligo repository for this:
[ticket](https://gitlab.com/ligolang/ligo/-/issues/1461).

Meanwhile, rewrite the failing test to avoid the flawed case, and add a
regression test (yet disabled) for LIGO logic that should hit the issue
- this test is to be enabled later.