[LIGO-802] Move WebIDEM to Common module
Problem: WebIDEM is not a type that is visible externally, so it should
not go into the Types module. We should have a separate module for it.

Solution: Create the Common module as the place for types that are used
across the backend (i.e. "common" to a lot of files in the backend), but
are not visible externally as part of the API.