[LIGO-218] Decode modules from get-scope
Problem: LIGO now outputs information about modules in `get-scope`. So
we should decode it there.

Solution: First, I needed to refactor `FromCompiler` a bit, so we now
can decode modules and module aliases with ease, in a similar manner to
variables. It will also help with LIGO-593 hopefully.

Of course, after this refactoring, actually decode from modules and
module aliases. There is a catch: `ScopedDecl` doesn't remember about
aliases, so I added `_mdsInit :: Module` which remembers whether it's a
module declaration (although not the decl contents) or whether it's an
alias. Also add a parser for it.

In `Cli.Json`, decode the new module fields and datatypes.