[LIGO-218] Resolve alias chains at decl
Problem: Resolving chains of aliases on the use site rather than during
the declaration has a problem: if we have something like `module A = A`
and `A` is in scope, then we'll create an infinite loop when trying to
use the alias, because it's going to always find this alias in-scope and
try to resolve it with itself.

Solution: Save the chain of `ExpandedDeclRef`s since the module
declaration. Note we insert `Nothing` if it fails to resolve the alias
at any point.