[LIGO-802] Refactor writing files for each request
Problem: Many requests requiring writing the entire project to disk
before calling ligo on the project. This code is duplicated in each
request.

Solution: Create a function `withProject` that writes the entire project
to disk, calls a monadic callback, and then cleans up the project. Then
use this function in the implementation of each request.