Handle `IOException` when calling `ligo`
Problem: in case `ligo` run fails, currently `IOException` will be
thrown.

This seems a bit inconvenient for two reasons:
* If I run a large block of code, I can't be sure where `IOException`
  fired from exactly, there might be multiple reasons for that.
* `callLigoImpl` already can throw numerious different exceptions, and I
  can't be sure that one day someone adds a dedicated exception for IO
  error. If I rely on `IOException` now, that change will break my code.

And in debugger code I would benefit from a way to handle `ligo` run
exceptions.

Solution: add a error handler to `callLigoImpl` that rethrows
`IOException` as a custom ligo exception.

At a first glance, language server has a lot of prior checks on that
`ligo` executable exists and so it won't benefit from this commit.
But still, it is always possible that `ligo` executable vanishes
right in the middle of the work, and I think that providing a better
error message for this case won't harm.