[LIGO-817] Make correct types from the LIGO table
Problem: In `AST.Scope.ScopedDecl.Parser`, we try to read from `TSum`
assuming there will be a list of `Variant`s, and these will consist of
`Ctor`s. However, when decoding from JSON, we build `TypeName`s inside
`TField`s, as if it was a record, even in the case of a sum. Because of
this, when parsing, we use `mapMaybe`, causing it to form an empty list,
which our pretty printers error when it's an empty list.

Solution: Add a `FieldKind` to keep track of whether we're decoding a
sum or a record. Now we can properly insert `Variant` vs `TField`, and
also `Ctor` vs `FieldName` vs `TypeName`, and our parsers will be happy
again.