[Chore] Get rid of decoding from type_content
Problem: It makes the output look ugly. For instance, the next commit
will add two hover tests in `sum.{js,m}ligo`. If you hovered over the
parameter with this case, it would completely expand its type to the
type definition. Without it, it will expand to simply `parameter`, which
is the type alias.

Solution: Get rid of this case.