[Chore] Fix `prepareRenameDeclaration`
Problem: We return the range of the given node, but the documentation
says we should return the range of its declaration. Also, this will mess
up tests in the next commit if not fixed.

Solution: Use `definitionOf` instead of `rangeOf`. Also, since `rangeOf`
was created exclusively for `prepareRenameDeclarationAt`, let's delete
it since it's unneeded now.