[LIGO-795] Handle bad entrypoint correctly
Problem: at this moment if I specify the wrong entrypoint in
`launch.json` then I'll get an error in the bottom right corner.
It's expected that I see a window with this error message and with
button `Open launch.json`.Moreover, I would see an input box
for parameter/storage but I couldn't write something
there because validation breakes.

Solution: make a contract metadata optional in `handleGetContractMetadata`
and return `Nothing` if an exception happens in `compileLigoContractDebug`.
Also, add an extra check for compiling a contract in `initDebuggerSession`