[LIGO-800] Extract validation to a separate class
Problem: currently, in the code for input boxes, the code for handling
validation is mixed up with the code of the input box, and it makes it
harder to bring some more complex solution for validation while staying
with a reasonably simple code.

Solution: extract the validation functionality to a separate class.

As far as I believe, this refactoring does not change the semantics of
the code. I only added a `try-catch` section for the event listener, as
errors there are not reported anyhow.