[LIGO-800] Avoid tasks stucking in validation
Problem: when entering parameter/storage into the input box, if I type a
lot of text quickly, I will wait for the validation for quite a long
time to pass.

This happens because our backend processes requests strictly
sequentially, and for every intermediate text in the input box we create
a validation task that backend processes at its time. Running that many
validation tasks takes long.

Solution: re-write validation at frontend so that we do not spawn so
many tasks. If some value is currently being validated, there is no
sense in keeping many values that wait for validation, rather we can
pick only the last one for the next validation.

So we keep running one validation at a time, and never abort it
(trully aborting the computation would be hard, and we don't want to
spawn many `ligo` processes in parallel for the sake of reducing
latency). Rather we keep the queue of values pending for validation
limited to the size of 1.

Comments: this solution is not perfect.

If running a validation takes `X`, then in the worst case the user will
wait for `2 * X` after stopping typing new characters.

However this problem seems unavoidable given that we refused to try to
abort async computations.