[LIGO-844] Use better name for pipe
Problem: we used to create a pipe for communication between the
extension and the backend, and assign a completely random name to it.

Now when I, for my personal needs, visit my `/tmp` folder, I get some
weird file names among fully clear filenames, and that litters my space.

Solution: add a prefix indicating what the file is.