[LIGO-768] Move `DebuggerException`
Problem: I want to use `DebuggerException` in many places in the
project, so better to have it at the very bottom of modules hierarchy.
Even lower than `CLI.Types` module because we don't want e.g.
`Michelson` module to import it.

Solution: move it to a separate `Error` module.

Now `SomeDebuggerException` must be moved somewhere else. It, on
contrary, should reside at the top of the modules hierarchy.
So I put it in `Handlers.Helpers`, since `Handlers.*` modules need to
catch those exceptions.