[LIGO-768] Use `Michelson.DecodeError` as exception
Problem: currently `DecodeError` in `Michleson` module we do not throw
as exception, rather turn it to string. This is not good.

Solution: add `MichelsonDecodeException` wrapper since our style guide
does not allow using the same type as error in pure context and as
exception.

Add `DebuggerException` instance for it. And do not turn it to string
rather throw as exception.