[LIGO-768] Replace `DapMessageException`
Problem: we use `DapMessageException` as a stub for every uncertain
error situation, and this is not good. It seems better to split it
further into separate exception types that handle a certain category of
cases each.

Solution: most of those cases are about bad `launch.json` being
provided, so I create `ConfigurationException`.

There is also one case where `DapMessageException` is used for an
unexpected case of communication between the plugin and the adapter, I
create a dedicated exception here too.