[LIGO-768] Make `parseValue` return `Either`
Problem: I can't help but find it weird that `parseValue` throws
exception in case of validation failure, and when calling it we use
`try @SomeDebuggerException` to get the error.

It shouldn't work so, at least because in `parseValue` we have two
sources of errors (`parseExpandValue` and `ligo` executable), and this
function knows better where to get the error message in each case.

Solution: make `parseValue` return `Either` instead of throwing
exceptions in case of parsing failure.