[LIGO-768] Add exception tags and use them in typescript part
Problem: at this moment when we send an error response its `format` field
isn't used in typescript part. Moreover, all exceptions that are coming
from adapter are treated in a similar way. This is not convenient because
sometimes we want to show a dialog message
(e.g. when we get an `UnsupportedLigoVersionException` we want to show
a button which opens a page with the latest supported LIGO version)

Solution: add `ExceptionTag` type and insert it in `response.message`
field. Then resolve `DAPErrorResponse` on typescript side properly