[LIGO-883] Add `getAvailableEntrypoints` check for each contract
Problem: in one of the LIGO's latest versions the command
`ligo info list-declarations` was broken for contracts with functions
from stdlib. We've discovered this when testing some contracts manually,
from tests everything seemed okay. We should add a check for this command
for each contract

Solution: run `getAvailableEntrypoints` for each contract in
`Contracts are sensible` test. As a bonus I made running these checks
in parallel