Fix weights of REPLACE
A REPLACE a BY b is equivalent to an INSERT a and DELETE b,
so its weight should be the sum of the INSERT and DELETE.

This is why the weight is now the sum of the "size" of
what's been deleted plus what's been added.

This fixes the a * b vs. b * c case, whose diff was wrongly :
- a
+ b
- b
+ c
and is now, correctly :
- a
  b
+ c