[LIGO-799] Rename all config resolution commands to one
Problem: now there is no sense in having commands like `AskForParameter`
and `AskForStorage`, we can use the same name everywhere and that will
be simpler for everyone.

Solution: do the replacement.

Also, in the documentation that I had to update - do some minor changes
to better reflect the new approach to variables-commands.