[LIGO-896] Separate MR templates
Problem: we have a default MR template set up in the project's settings
in GitLab, however for Debugger it is now what we want. In particular,
the point about syntax hightlighting is not related to Debugger.

Solution: add our MR template to `.gitlab/merge_request_templates`.
Put the previous default template a template file too.
The previous default template in the projects settings - replace it with
a hint to select the template.