[LIGO-729] Add tests for finding type names
Problem: There are a few commented tests about LIGO-700 and LIGO-729.

Solution: For Find tests, simply uncomment them. For Highlights test, I
actually refactored them a bit to avoid opening and closing the same
document all the time, and to avoid repetition of the tests.

Note: One of the tests (for LIGO-700) was wrong, it was expecting a
reference for the 2nd commit, but it's using the 1st commit, so I
changed it.