[LIGO-582] Strip duplicate locations at `readLigoMapper` stage
Problem: at this moment if we have snapshots that point to the same location
then we just leave the last one. However, we discovered some problems
with this approach. E.g. we have a cycle with only one location meta.
In this case we'll leave only the last iteration. Moreover, if this
cycle is infinite then execution would halt.

Solution: strip duplicate locations at `readLigoMapper` stage.
Semantically we do the same thing but instead of stripping snapshots
we're replacing duplicate location metas with `Nothing`