[LIGO-768] Change typecheck error in `parseValue`
Problem: currently, if typechecking fails, we report the entire
typechecking error which is multiline, contains some diff, and repeats
the typechecked value.

When such a message appears in the input box for parameter/storage, that
is hard to read.

Solution: ignore the actual typechecking message, instead provide our
own that only mentiones that typechecking failed, and the expected type
in Michelson format. Michelson - for now, switching to LIGO is left for
the dedicated ticket as currently LIGO provides no way of obtaining a
type.